LayoutTests/imported/w3c/web-platform-tests/content-security-policy/frame-src/frame-src-redirect-expected.txt

CONSOLE MESSAGE: The Content Security Policy 'frame-src http://foo.test' was delivered in report-only mode, but does not specify a 'report-uri'; the policy will have no effect. Please either add a 'report-uri' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: The Content Security Policy 'frame-src http://foo.test' was delivered in report-only mode, but does not specify a 'report-uri'; the policy will have no effect. Please either add a 'report-uri' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: [Report Only] Refused to load http://localhost:8800/common/redirect.py?location=http%3A%2F%2Fxn--n8j6ds53lwwkrqhv28a.localhost%3A8800%2Fcontent-security-policy%2Fframe-src%2Fsupport%2Fframe.html because it does not appear in the frame-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/content-security-policy/frame-src/support/frame.html because it does not appear in the frame-src directive of the Content Security Policy.
CONSOLE MESSAGE: [Report Only] Refused to load http://xn--n8j6ds53lwwkrqhv28a.localhost:8800/content-security-policy/frame-src/support/frame.html because it does not appear in the frame-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

TIMEOUT Redirected iframe src should evaluate both enforced and report-only policies on both original request and when following redirect Test timed out
